==> 4-XGBoost_v2/1-Prep_data_Lung.Rmd <==
---
title: "Generate peptide database"
author: "Kaspar Bresser"
date: "22/03/2024"

output: 
  github_document:
    toc: true
  html_document: 
    theme: simplex
    highlight: pygments
#    code_folding: show
    self_contained: TRUE
    toc: yes
    toc_float:
      collapsed: no
---



```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      message=FALSE,
                      eval = F,
                      autodep  = TRUE,
                      cache = FALSE,
#                     comment = NA,
                      fig.width = 5,
                      fig.asp = 0.618,
                      fig.align = "center")
```

```{r startup, eval =TRUE}
library(babelgene)
library(tidyverse)
library(furrr)
```

# Lung data peptide select

## Import ligands


```{r check_sheets, eval =TRUE}
MS.data <- read_tsv("Data/Lung_HLA_MS_data.tsv")
```



```{r import_ligand_data, eval =TRUE}
tumors <- c("C3N-02287",
            "C3N-02289",
           "C3N-02671",
           "C3N-03421")

MS.data %>% 
  select(PEP.GroupingKey, PG.Genes, contains(tumors)) %>% 
  pivot_longer(-c(PEP.GroupingKey, PG.Genes)) %>% 
  mutate(tumor = str_split_i(name, "_", 1)) %>% 
  na.omit() %>% 
  distinct(PEP.GroupingKey, tumor, .keep_all = T) -> MS.data

MS.data %>% count(tumor)

```





Tidy up, and add peptide lengths.

```{r tidy_broad_data, eval =TRUE}
MS.data %>% 
  transmute(sequence = PEP.GroupingKey, tumor = tumor, gene.symbol = PG.Genes) %>% 
  mutate(pep.length = str_length(sequence))  -> MS.data

MS.data
```

Check the amount of 9mers for each allele. 

```{r check_9mers, eval =TRUE}
pep.lengths <- count(MS.data, tumor, pep.length)

pep.lengths %>% 
  filter(pep.length %in% 8:12) %>% 
ggplot( aes(x = pep.length, y = n, color = tumor))+
  geom_line()
```


```{r, eval =TRUE}
babelgene::orthologs(genes = MS.data$gene.symbol, species = "mouse", top = T) %>% 
  transmute(gene.symbol = human_symbol, ensemble_id =  human_ensembl) %>% 
  full_join(read_tsv("Data/conversion.txt"), by = c("ensemble_id" = "Ensembl Gene ID")) %>% 
  as_tibble() %>% 
  na.omit() %>% 
  distinct(gene.symbol, `UniProt/SwissProt Accession`, .keep_all = T) %>% 
  transmute(gene.symbol = gene.symbol, swissprot.id = `UniProt/SwissProt Accession`) -> conversion.table
```


```{r}
read_tsv("Data/Lung_Test_Table_all.tsv") %>% 
  filter(ligand == TRUE) %>% 
  filter(tumor %in% tumors) %>% 
  select(sequence, tumor) -> peptides.in.test
```


Subsample the peptide pools to get a more workable number, 500 is slightly under the smallest allele-set. 

```{r subsample}
"../2-random_forest_analyses/Data/Protein_per_Uniprot_entry_library_v2_RBP_GC_length_codon_AA_m6A_m5C_AtoI_m1A_m7G_CD8miRDB_PTM.csv" %>% 
  read_tsv() %>% 
  pull(Entry) -> IDs.in.library

MS.data %>% 
  count(tumor, pep.length)

MS.data %>% 
  filter(pep.length %in% 9:11) %>% 
  left_join(conversion.table) %>% 
  filter(swissprot.id %in% IDs.in.library) %>% 
  distinct(tumor, sequence, .keep_all = T) %>% 
  anti_join(peptides.in.test, by = c("tumor", "sequence")) %>% 
  group_by(tumor) %>% 
  slice_sample(n = 10000) %>% 
  mutate(ligand = TRUE) %>% 
  ungroup() -> filtered.peptides


write_tsv(filtered.peptides, "Output/Lung_peptides_Train.tsv")
```

## Select decoys

Import UniProt sequences, use the rna expression table to make an ensembl/UniProt matching table.

Define a function that, for each allele: (1) Checks which sequences were detected for that allele, (2) samples an excess of proteins from the uniprot table weighted by their length, (3) then for each sample extracts a 9mer, (4) Remove detected peptides, filter for uniqueness and down sample to 350,000. 


```{r import_uniprot}

uniprot <- read_csv("Data/UniProt_reviewed_input.tsv") %>% rename(swissprot.id = "sequence_id")


get_peptide <- function(tum, pep.len){
  MS.data %>% 
    mutate(sequence = str_to_upper(sequence)) %>% 
    filter(tumor == tum) %>% 
    filter(nchar(sequence) == pep.len) %>% 
    pull(sequence) %>% 
    toupper() -> s
  
    amount.needed <- nrow(filter(filtered.peptides, tumor == tum & nchar(sequence) == pep.len))*10
  
  uniprot %>% 
    mutate(len = nchar(sequence)) %>% 
    slice_sample(n = amount.needed*2, weight_by = len, replace = T) %>% 
    rowwise() %>% 
    mutate(number = sample(1:(nchar(sequence)-pep.len), 1),
         sequence = str_sub(sequence, number, number+(pep.len-1))) %>% 
    ungroup() %>% 
    filter(!(sequence %in% s)) %>% 
    transmute(sequence = sequence, swissprot.id = swissprot.id, tumor = tum, ligand = FALSE) %>% 
    filter(swissprot.id %in% IDs.in.library) %>% 
    distinct(sequence, .keep_all = T) %>% 
    slice_sample(n = amount.needed)
}

decoys <- map2_dfr(rep(tumors, 3), rep(c(9,10,11), each = 4), get_peptide)

filtered.peptides %>% 
  select(!c(gene.symbol, pep.length)) %>% 
  bind_rows(decoys) -> test.set


test.set %>% 
  count(tumor, ligand, nchar(sequence))

```

Write out the table and peptide files

```{r, echo=FALSE}
test.set %>% 
  transmute(pep.len = paste0(nchar(sequence), "AA"), sequence = sequence, tumor = tumor) %>% 
  unite("info", tumor, pep.len) %>% 
  nest(data = sequence) %>% 
  mutate(peptides = map(data, pull, sequence),
         peptides = set_names(peptides, info)) %>% 
  pull(peptides) %>% 
  map2(names(.), ~write_lines(.x, paste0("./Output/Lung/peptides/Lung_Test_Peptides_", .y, ".tsv")))

write_tsv(test.set, "./Output/Lung/Lung_Test_Table.tsv")
```


## Run netMHCpan

Get the file names

```{r}
files <- list.files("Output/Lung/peptides", pattern = "Test_Pep", full.names = T)

files
```


```{r}
HLA.alleles <- list(  `C3N-02287` =	c('A0201','A0301','B3801','B4405','C0202','C1203'),
                      `C3N-02289` =	c('A0301','A2601','B1801','C0701','C1203'),
                      `C3N-02671` =	c('A0101','A0201','B0801','B4402','C0501','C0701'),
                      `C3N-03421` =	c('A0101','A0201','B0801','C0202','C0701'))
```


Run system command for netMHCpan, for each allele. Define function to get the arguments/flags for netMHCpan


```{r}
args_netMHC <- function(allele, file.name, peptide.len){
  c(
    paste0("-a HLA-", sub("(.{3})(.*)", "\\1:\\2", allele)),
    paste0("-f ", file.name),
    "-p",
    "-rth 0.0",
    "-rlt 0.0",
    paste0("-l ", peptide.len),
    "-t -100"
  )
}

```

wrapper function to predict for each length

```{r}
netMHC_wrapper <- function(tum, len){
  
  file.name <- paste0("Output/Lung/peptides/Lung_Test_Peptides_", tum, "_", len, "AA.tsv")
  all.alleles <- HLA.alleles[[tum]]
  
  future_map(all.alleles, ~system2(command = "netMHCpan", 
                      args = args_netMHC(., file.name, len), 
                      stdout  = paste0("./Output/Lung/netMHCpan_predictions/netMHCpan_prediction_",tum, "_", .,
                                      "_", len , "AA_yay.txt")))
}
```

Run netMHCpan, parallel for the alleles 

```{r}
plan(multisession, workers = 6)


map2(rep(tumors, 3), rep(9:11, each = 4), ~netMHC_wrapper(.x, .y))
```


### Cleanup and combine per tumor

Define function to read in predictions for each peptide length, cleanup the netMHC output, and combine them in a single tsv file

```{r}
combine_and_clean <- function(tum){

  file.list <- list.files("./Output/Lung/netMHCpan_predictions/", pattern = tum, full.names = T)
#  file.list <- paste0("./Output/netMHCpan_predictions/", file.list)
  
  file.list %>% 
    map(read_lines) %>% 
    map(~.[grepl("   1 HLA", .)]) %>% 
    unlist() %>% 
    c(" Pos         MHC        Peptide      Core Of Gp Gl Ip Il        Icore        Identity  Score_EL %Rank_EL", .) %>% 
    write_lines(paste0("./Output/Lung/netMHCpan_predictions_clean/", tum,".tsv"))
  
  gc()
}
```


Apply the function for each allele

```{r}

map(tumors, combine_and_clean)
```


### Finalize files

Write a function that reads in the allele predictions per tumor line, only imports the columns needed, joins them together, and selects the highest affinity rank.

```{r}
write_tumor_files <- function(fil){
 tum <- str_split_i(fil, "\\/|\\.", 4)
  fil %>% 
    read_table() %>% 
    distinct(Peptide, MHC, .keep_all = T) %>% 
    transmute(allele = str_remove_all(MHC, "HLA-|\\*|\\:"), sequence = Peptide, rank = `%Rank_EL`) %>% 
    group_by(sequence) %>% 
    slice_min(rank, n=1) %>% 
    write_tsv(paste0("./Output/Lung/netMHCpan_predictions_clean/HLA_affinities_",tum,".tsv"))
}
```

Define HLA typings per tumor line and run the function.

```{r}
list.files("Output/Lung/netMHCpan_predictions_clean", full.names = T)
map(list.files("Output/Lung/netMHCpan_predictions_clean", full.names = T), write_tumor_files)
```





## Run netMHCpanExp

Get the file names

```{r}
files <- list.files("Output/Lung/peptides/", pattern = "Test_Pep", full.names = T)

files
```


```{r}
HLA.alleles 
```


Run system command for netMHCpan, for each allele. Define function to get the arguments/flags for netMHCpan


```{r}
args_netMHC <- function(allele, file.name, peptide.len){
  c(
    paste0("-a HLA-", sub("(.{3})(.*)", "\\1:\\2", allele)),
    paste0("-f ", file.name),
    "-p",
    "-rth 0.0",
    "-rlt 0.0",
    paste0("-l ", peptide.len),
    "-t -100",
    "-inpfmt 0"
  )
}

```

wrapper function to predict for each length

```{r}
netMHC_wrapper <- function(tum, len){
  
  file.name <- paste0("Output/Lung/peptides/Lung_Test_Peptides_", tum, "_", len, "AA.tsv")
  all.alleles <- HLA.alleles[[tum]]
  
  future_map(all.alleles, ~system2(command = "netMHCpanExp", 
                      args = args_netMHC(., file.name, len), 
                      stdout  = paste0("./Output/Lung/netMHCpanExp_predictions/netMHCpan_prediction_",tum, "_", .,
                                      "_", len , "AA_yay.txt")))
}
```

Run netMHCpan, parallel for the alleles 

```{r}
plan(multisession, workers = 6)


map2(rep(tumors, 3), rep(9:11, each = 4), ~netMHC_wrapper(.x, .y))
```


### Cleanup and combine per tumor

Define function to read in predictions for each peptide length, cleanup the netMHC output, and combine them in a single tsv file

```{r}
combine_and_clean <- function(tum){

  file.list <- list.files("./Output/Lung/netMHCpanExp_predictions/", pattern = tum, full.names = T)
#  file.list <- paste0("./Output/netMHCpan_predictions/", file.list)
  
  file.list %>% 
    map(read_lines) %>% 
    map(~.[grepl("   1 HLA", .)]) %>% 
    unlist() %>% 
    c(" Pos         MHC        Peptide      Core Of Gp Gl Ip Il        Icore        Identity  Score_EL %Rank_EL Exp", .) %>% 
    write_lines(paste0("./Output/Lung/netMHCpanExp_predictions_clean/", tum,".tsv"))
  
  gc()
}
```


Apply the function for each allele

```{r}

map(tumors, combine_and_clean)
```


### Finalize files

Write a function that reads in the allele predictions per tumor line, only imports the columns needed, joins them together, and selects the highest affinity rank.

```{r}
write_tumor_files <- function(fil){
 tum <- str_split_i(fil, "\\/|\\.", 4)
  fil %>% 
    read_table() %>% 
    distinct(Peptide, MHC, .keep_all = T) %>% 
    transmute(allele = str_remove_all(MHC, "HLA-|\\*|\\:"), sequence = Peptide, rank = `%Rank_EL`) %>% 
    group_by(sequence) %>% 
    slice_min(rank, n=1) %>% 
    write_tsv(paste0("./Output/Lung/netMHCpanExp_predictions_clean/HLA_affinities_",tum,".tsv"))
}
```

Define HLA typings per tumor line and run the function.

```{r}
list.files("Output/Lung/netMHCpanExp_predictions_clean", full.names = T)
map(list.files("Output/Lung/netMHCpanExp_predictions_clean", full.names = T), write_tumor_files)
```






## Final combine data

```{r}
train.table <- read_tsv("Output/Lung/Lung_Train_Table.tsv")
```



```{r}
files <- list.files("Output/Lung/netMHCpan_predictions_clean", pattern = "HLA", full.names = T)

files %>% 
  map(read_tsv) %>% 
  set_names(str_split_i(files, "\\_|\\.", 5)) %>% 
  map2(names(.), ~mutate(.x, tumor = .y)) %>% 
  reduce(bind_rows) %>% 
  mutate(tumor = as.character(tumor)) %>% 
  select(-allele) -> netMHCranks
```




```{r}
files <- list.files("Output/Lung/netMHCpanExp_predictions_clean", pattern = "HLA", full.names = T)

files %>% 
  map(read_tsv) %>% 
  set_names(str_split_i(files, "\\_|\\.", 5)) %>% 
  map2(names(.), ~mutate(.x, tumor = .y)) %>% 
  map( rename, rankExp = rank) %>% 
  reduce(bind_rows) %>% 
  mutate(tumor = as.character(tumor)) %>% 
  select(-allele) -> netMHCranksExp
```


```{r}
train.table %>% 
  mutate(tumor = as.character(tumor)) %>% 
  left_join(netMHCranks, by = c("sequence", "tumor")) %>% 
  left_join(netMHCranksExp, by = c("sequence", "tumor")) %>% 
  distinct(sequence, tumor, .keep_all = T) %>% 
  na.omit() -> train.table

train.table

write_tsv(train.table, "Output/Lung_Train_Table_all.tsv")
```
